Fix: Resolve encoding problems in copyright annotation
To fix encoding problems in the copyright annotation in the header commmits in `cook-and-pack.cmd`.
When lines end with Chinese characters, we now append a space character at the end of those lines.

Changes Made:
 - Added a space character at the end of lines that end with Chinese characters in
    `cook-and-pack.cmd` - not only commits.

Reverts commit 3c717028.

The commit 3c717028 attempted to solve the problem but was unsuccessful. We have now reversed it
and replaced it with a new solution.